pyproject.toml
[project]
name = "chiffrement-hybride"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pycryptodome>=3.23.0",
]
